fix(delete album infra): fix delete infra throw error logic

+++ src/application/useCases/album/DeleteAlbum.ts
@@ -1,11 +1,14 @@
 import type { AlbumDeleteInputDTO } from '#application/DTOs/album/AlbumDeleteInputDTO.js'
 import type { AlbumDeleteOutputDTO } from '#application/DTOs/album/AlbumDeleteOutputDTO.js'
+import { AlbumNotFoundError } from '#application/errors/AlbumNotFoundError.js'
 import type { AlbumRepository } from '#application/ports/AlbumRepository.js'
 
 export class DeleteAlbumUseCase {
   constructor(private readonly albumRepo: AlbumRepository) {}
 
-  async execute(dto: AlbumDeleteInputDTO): Promise<AlbumDeleteOutputDTO> {
-    await this.albumRepo.delete(dto.id)
+	async execute(dto: AlbumDeleteInputDTO): Promise<AlbumDeleteOutputDTO> {
+		const exists = await this.albumRepo.findById(dto.id)
+		if (!exists) throw ne AlbumNotFoundError
+    	await this.albumRepo.delete(dto.id)
   }
 }
